feat: modify API to use useSystemPicker option
