+++ lib/browser/api/session.ts
@@ -1,24 +1,29 @@
 import { fetchWithSession } from '@electron/internal/browser/api/net-fetch';
-
-import { net } from 'electron/main';
-
+import { desktopCapturer, net } from 'electron/main';
 const { fromPartition, fromPath, Session } = process._linkedBinding('electron_browser_session');
 const { isDisplayMediaSystemPickerAvailable } = process._linkedBinding('electron_browser_desktop_capturer');
 
-// Fake video window that activates the native system picker
-// This is used to get around the need for a screen/window
-// id in Chrome's desktopCapturer.
-let fakeVideoWindowId = -1;
-// See content/public/browser/desktop_media_id.h
-const kMacOsNativePickerId = -4;
-const systemPickerVideoSource = Object.create(null);
-Object.defineProperty(systemPickerVideoSource, 'id', {
-  get () {
-    return `window:${kMacOsNativePickerId}:${fakeVideoWindowId--}`;
+async function getNativePickerSource () {
+  if (process.platform !== 'darwin') {
+    throw new Error('Native system picker option is currently only supported on MacOS');
   }
-});
-systemPickerVideoSource.name = '';
-Object.freeze(systemPickerVideoSource);
+
+  if (!isDisplayMediaSystemPickerAvailable) {
+    throw new Error(`Native system picker unavailable. 
+      Note: This is an experimental API; please check the API documentation for updated restrictions`);
+  }
+
+  // Pass in the needed options for a more native experience
+  // screen & windows by default, no thumbnails, since the native picker doesn't return them
+  const options: Electron.SourcesOptions = {
+    types: ['screen', 'window'],
+    thumbnailSize: { width: 0, height: 0 },
+    fetchWindowIcons: false
+  };
+
+  const mediaStreams = await desktopCapturer.getSources(options);
+  return mediaStreams[0];
+}
 
 Session.prototype.fetch = function (input: RequestInfo, init?: RequestInit) {
   return fetchWithSession(input, init, this, net.request);
@@ -29,7 +34,7 @@ Session.prototype.setDisplayMediaRequestHandler = function (handler, opts) {
 
   this._setDisplayMediaRequestHandler(async (req, callback) => {
     if (opts && opts.useSystemPicker && isDisplayMediaSystemPickerAvailable()) {
-      return callback({ video: systemPickerVideoSource });
+      return callback({ video: await getNativePickerSource() });
     }
 
     return handler(req, callback);
